tcp server init

+++ rocket/net/io_thread_group.cc
@@ -7,7 +7,7 @@ namespace rocket {
 
 IOThreadGroup::IOThreadGroup(int size) : m_size(size) {
   m_io_thread_groups.resize(size);
-  for (size_t i = 0; i < size; ++i) {
+  for (size_t i = 0; (int)i < size; ++i) {
     m_io_thread_groups[i] = new IOThread();
   }
 }
@@ -29,7 +29,7 @@ void IOThreadGroup::join() {
 } 
 
 IOThread* IOThreadGroup::getIOThread() {
-  if (m_index == m_io_thread_groups.size() || m_index == -1)  {
+  if (m_index == (int)m_io_thread_groups.size() || m_index == -1)  {
     m_index = 0;
   }
   return m_io_thread_groups[m_index++];
